requirements.txt
camel-ai
python-dotenv
pdfplumber
streamlit
